adding checkpoint #223

+++ devnet/docker/icon-bsc/scripts/token.solidity.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
 source utils.sh
-# Parts of this code is adapted from https://github.com/icon-project/btp/blob/goloop2moonbeam/testnet/goloop2moonbeam/scripts
 
 eth_blocknumber() {
   curl -s -X POST $BSC_RPC_URI --header 'Content-Type: application/json' \
@@ -8,50 +7,90 @@ eth_blocknumber() {
 }
 
 deploy_solidity_bmc() {
-  echo "deploying solidity bmc"
   cd $CONTRACTS_DIR/solidity/bmc
-  cp $ICONBRIDGE_BASE_DIR/bin/env ./.env
-  rm -rf contracts/test build .openzeppelin
-  truffle compile --all
-  BMC_BTP_NET=$BSC_BMC_NET \
-    truffle migrate --network bsc --compile-all
-  eth_blocknumber > $CONFIG_DIR/btp.bsc.block.height
-  generate_metadata "BMC"
+  if [ ! -f $CONFIG_DIR/bsc.deploy.bmc ]; then
+    rm -rf contracts/test build .openzeppelin
+    truffle compile --all
+    echo "deploying solidity bmc"
+    set +e
+    for i in $(seq 1 20); do
+      BMC_BTP_NET=$BSC_BMC_NET \
+      truffle migrate --network bsc --compile-none
+      if [ $? == 0 ]; then
+          break
+      fi
+      echo "Retry: "$i
+    done
+    set -e
+    eth_blocknumber > $CONFIG_DIR/bsc.chain.height
+    generate_metadata "BMC"
+    echo -n "bmc" > $CONFIG_DIR/bsc.deploy.bmc
+  fi
 } 
 
 deploy_solidity_bts() {
   echo "deploying solidity bts"
   cd $CONTRACTS_DIR/solidity/bts
-  cp $ICONBRIDGE_BASE_DIR/bin/env ./.env
-  rm -rf contracts/test build .openzeppelin
-  truffle compile --all
-  BSH_COIN_NAME="BNB" \
-  BSH_COIN_FEE=100 \
-  BSH_FIXED_FEE=5000 \
-  BMC_PERIPHERY_ADDRESS="$(cat $CONFIG_DIR/btp.bsc.bmc.periphery)" \
-  truffle migrate --compile-all --network bsc --f 1 --to 1
-  generate_metadata "BTS"
-}
+  if [ ! -f $CONFIG_DIR/bsc.deploy.bts ]; then
+    rm -rf contracts/test build .openzeppelin
+    truffle compile --all
+    set +e
+    for i in $(seq 1 20); do
+      BSH_COIN_NAME="BNB" \
+      BSH_COIN_FEE=100 \
+      BSH_FIXED_FEE=5000 \
+      BMC_PERIPHERY_ADDRESS="$(cat $CONFIG_DIR/bsc.addr.bmcperiphery)" \
+      truffle migrate --compile-none --network bsc --f 1 --to 1
+      if [ $? == 0 ]; then
+        break
+      fi
+      echo "Retry: "$i
+    done
+    set -e
+    generate_metadata "BTS"
+    echo -n "bts" > $CONFIG_DIR/bsc.deploy.bts
+  fi
+} 
 
 deploy_solidity_token() {
   echo "deploying solidity token " $1
   cd $CONTRACTS_DIR/solidity/bts
-  cp $ICONBRIDGE_BASE_DIR/bin/env ./.env
-  export BSH_COIN_NAME=$1
-  export BSH_COIN_SYMBOL=$2
-  export BSH_DECIMALS=18
-  export BSH_INITIAL_SUPPLY=100000
-  truffle migrate --network bsc --f 3 --to 3
-  jq -r '.networks[] | .address' build/contracts/ERC20TKN.json >$CONFIG_DIR/btp.bsc.$1
-  wait_for_file $CONFIG_DIR/btp.bsc.$1
+  if [ ! -f $CONFIG_DIR/bsc.deploy.coin$1 ]; then
+    set +e
+    for i in $(seq 1 20); do
+      BSH_COIN_NAME=$1 \
+      BSH_COIN_SYMBOL=$2 \
+      BSH_DECIMALS=18 \
+      BSH_INITIAL_SUPPLY=100000 \
+      truffle migrate --compile-none --network bsc --f 3 --to 3
+      if [ $? == 0 ]; then
+        break
+      fi
+      echo "Retry: "$i
+    done
+    set -e
+    jq -r '.networks[] | .address' build/contracts/ERC20TKN.json >$CONFIG_DIR/bsc.addr.$1
+    wait_for_file $CONFIG_DIR/bsc.addr.$1
+    echo -n $1 > $CONFIG_DIR/bsc.deploy.coin$1
+  fi
 }
 
 configure_solidity_add_bts_service() {
   echo "adding bts service into BMC"
   cd $CONTRACTS_DIR/solidity/bmc
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
-    --method addService --name "bts" --addr $(cat $CONFIG_DIR/btp.bsc.bts.periphery))
-  echo "$tx" >$CONFIG_DIR/tx/addService.bsc
+  if [ ! -f $CONFIG_DIR/bsc.configure.addbts ]; then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
+      --method addService --name "bts" --addr $(cat $CONFIG_DIR/bsc.addr.btsperiphery))
+    echo "$tx" >$CONFIG_DIR/tx/addService.bsc
+    isTrue=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$isTrue" == "1" ];
+    then
+      echo "addedBTS" > $CONFIG_DIR/bsc.configure.addbts
+    else
+      echo "Error Addding BTS"
+      return 1
+    fi  
+  fi
 }
  
 configure_solidity_add_bmc_owner() {
@@ -59,8 +98,18 @@ configure_solidity_add_bmc_owner() {
   BSC_BMC_USER=$(cat $CONFIG_DIR/keystore/bsc.bmc.wallet.json | jq -r .address)
   cd $CONTRACTS_DIR/solidity/bmc
   tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
-    --method addOwner --addr "0x${BSC_BMC_USER}")
-  echo "$tx" >$CONFIG_DIR/tx/addBmcUser.bsc
+    --method isOwner --addr "0x${BSC_BMC_USER}")
+  ownerExists=$(echo "$tx" | grep "IsOwner: true" | wc -l)
+  if [ "$ownerExists" == "0" ];then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
+      --method addOwner --addr "0x${BSC_BMC_USER}")
+    echo "$tx" >$CONFIG_DIR/tx/addBmcUser.bsc
+    ownerAdded=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$ownerAdded" != "1" ]; then
+      echo "Error adding bmc owner"
+      return 1 
+    fi
+  fi
 }
  
  configure_solidity_add_bts_owner() {
@@ -68,63 +117,134 @@ configure_solidity_add_bmc_owner() {
   BSC_BTS_USER=$(cat $CONFIG_DIR/keystore/bsc.bts.wallet.json | jq -r .address)
   cd $CONTRACTS_DIR/solidity/bts
   tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
-    --method addOwner --addr "0x${BSC_BTS_USER}")
-  echo "$tx" >$CONFIG_DIR/tx/addBtsUser.bsc
+    --method isOwner --addr "0x${BSC_BTS_USER}")
+  ownerExists=$(echo "$tx" | grep "IsOwner: true" | wc -l)
+  if [ "$ownerExists" == "0" ];then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
+      --method addOwner --addr "0x${BSC_BTS_USER}")
+    echo "$tx" >$CONFIG_DIR/tx/addBtsUser.bsc
+    ownerAdded=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$ownerAdded" != "1" ]; then
+      echo "Error adding bts owner"
+      return 1
+    fi  
+  fi
 }
 
 configure_solidity_set_fee_ratio() {
   echo "SetFee Ratio"
   cd $CONTRACTS_DIR/solidity/bts
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
-    --method setFeeRatio --name BNB --feeNumerator 100 --fixedFee 5000)
-  echo "$tx" >$CONFIG_DIR/tx/setFee.bsc
+  if [ ! -f $CONFIG_DIR/bsc.configure.setfee ]; then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
+      --method setFeeRatio --name BNB --feeNumerator 100 --fixedFee 5000)
+    echo "$tx" >$CONFIG_DIR/tx/setFee.bsc
+    status=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$status" != "1" ]; 
+    then
+      echo "Error setting fee ratio"
+      return 1
+    else
+      echo "feeSet" > $CONFIG_DIR/bsc.configure.setfee
+    fi  
+  fi
 }
  
 add_icon_link() {
-  echo "adding icon link $(cat $CONFIG_DIR/btp.icon.btp.address)"
+  echo "adding icon link $(cat $CONFIG_DIR/icon.addr.bmcbtp)"
   cd $CONTRACTS_DIR/solidity/bmc
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
-    --method addLink --link $(cat $CONFIG_DIR/btp.icon.btp.address) --blockInterval 3000 --maxAggregation 2 --delayLimit 3)
-  echo "$tx" >$CONFIG_DIR/tx/addLink.bsc
+  if [ ! -f $CONFIG_DIR/bsc.configure.addLink ]; then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
+    --method addLink --link $(cat $CONFIG_DIR/icon.addr.bmcbtp) --blockInterval 3000 --maxAggregation 2 --delayLimit 3)
+    echo "$tx" >$CONFIG_DIR/tx/addLink.bsc
+    status=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$status" != "1" ]; 
+    then
+      echo "Error adding link"
+      return 1
+    else
+      echo "addedLink" > $CONFIG_DIR/bsc.configure.addLink
+    fi 
+  fi
 }
 
 set_link_height() {
   echo "set link height"
   cd $CONTRACTS_DIR/solidity/bmc
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
-    --method setLinkRxHeight --link $(cat $CONFIG_DIR/btp.icon.btp.address) --height $(cat $CONFIG_DIR/btp.icon.block.height))
-  echo "$tx" >$CONFIG_DIR/tx/setLinkRxHeight.bsc
+  if [ ! -f $CONFIG_DIR/bsc.configure.setLink ]; then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
+    --method setLinkRxHeight --link $(cat $CONFIG_DIR/icon.addr.bmcbtp) --height $(cat $CONFIG_DIR/icon.chain.height))
+    echo "$tx" >$CONFIG_DIR/tx/setLinkRxHeight.bsc
+      status=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$status" != "1" ]; 
+    then
+      echo "Error setting link"
+      return 1
+    else
+      echo "setLink" > $CONFIG_DIR/bsc.configure.setLink
+    fi 
+  fi
 }
 
 add_icon_relay() {
   echo "adding icon relay"
   BSC_RELAY_USER=$(cat $CONFIG_DIR/keystore/bsc.bmr.wallet.json | jq -r .address)
   cd $CONTRACTS_DIR/solidity/bmc
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
-    --method addRelay --link $(cat $CONFIG_DIR/btp.icon.btp.address) --addr "0x${BSC_RELAY_USER}")
-  echo "$tx" >$CONFIG_DIR/tx/addRelay.bsc
+  if [ ! -f $CONFIG_DIR/bsc.configure.addRelay ]; then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
+      --method addRelay --link $(cat $CONFIG_DIR/icon.addr.bmcbtp) --addr "0x${BSC_RELAY_USER}")
+    echo "$tx" >$CONFIG_DIR/tx/addRelay.bsc
+    status=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$status" != "1" ]; 
+    then
+      echo "Error adding relay"
+      return 1
+    else
+      echo "addedRelay" > $CONFIG_DIR/bsc.configure.addRelay
+    fi 
+  fi
 } 
 
 
 bsc_register_wrapped_coin() {
   echo "bts: Register Wrapped Coin " $1
-  local btp_bts_fee_numerator=100
-  local btp_bts_fixed_fee=5000
+  local bts_fee_numerator=100
+  local bts_fixed_fee=5000
   cd $CONTRACTS_DIR/solidity/bts
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
-    --method register --name "$1" --symbol "$2" --decimals 18 --addr "0x0000000000000000000000000000000000000000" --feeNumerator ${btp_bts_fee_numerator} --fixedFee ${btp_bts_fixed_fee})
-  echo "$tx" >$CONFIG_DIR/tx/register.$1.bsc
+  if [ ! -f $CONFIG_DIR/bsc.register.coin$1 ]; then
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
+      --method register --name "$1" --symbol "$2" --decimals 18 --addr "0x0000000000000000000000000000000000000000" \
+      --feeNumerator ${bts_fee_numerator} --fixedFee ${bts_fixed_fee})
+    echo "$tx" >$CONFIG_DIR/tx/register.$1.bsc
+    status=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$status" != "1" ]; 
+    then
+      echo "Error registering wrapped coin " $1
+      return 1
+    else
+      echo "registered "$1 > $CONFIG_DIR/bsc.register.coin$1
+    fi 
+  fi
 }
 
 bsc_register_native_token() {
-  echo "bts: Register NativeCoin " $1
-  local btp_bts_fee_numerator=100
-  local btp_bts_fixed_fee=5000
-  local addr=$(cat $CONFIG_DIR/btp.bsc.$1) 
+  local bts_fee_numerator=100
+  local bts_fixed_fee=5000
+  local addr=$(cat $CONFIG_DIR/bsc.addr.$1) 
   cd $CONTRACTS_DIR/solidity/bts
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
-    --method register --name "$1" --symbol "$2" --decimals 18 --addr $addr --feeNumerator $btp_bts_fee_numerator --fixedFee ${btp_bts_fixed_fee})
-  echo "$tx" >$CONFIG_DIR/tx/register.$1.bsc
+  if [ ! -f $CONFIG_DIR/bsc.register.coin$1 ]; then
+    echo "bts: Register NativeCoin " $1
+    tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
+      --method register --name "$1" --symbol "$2" --decimals 18 --addr $addr --feeNumerator $bts_fee_numerator --fixedFee ${bts_fixed_fee})
+    echo "$tx" >$CONFIG_DIR/tx/register.$1.bsc
+    status=$(echo "$tx" | grep "status: true" | wc -l)
+    if [ "$status" != "1" ]; 
+    then
+      echo "Error registering native token " $1
+      return 1
+    else
+      echo "registered "$1 > $CONFIG_DIR/bsc.register.coin$1
+    fi 
+  fi
 }
 
 get_coinID() {
@@ -132,82 +252,15 @@ get_coinID() {
   cd $CONTRACTS_DIR/solidity/bts
   tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bts.js \
     --method coinId --coinName "$1")
-  echo "$tx" >$CONFIG_DIR/tx/coinID.$1
-}
-
-
-bsc_updateRxSeq() {
-  cd $CONTRACTS_DIR/solidity/bmc
-  truffle exec --network bsc "$SCRIPTS_DIR"/bmc.js \
-    --method updateRxSeq --link $(cat $CONFIG_DIR/btp.icon) --value 1
-}
-
-token_bsc_fundBSH() {
-  echo "Funding solidity BSH"
-  cd $CONTRACTS_DIR/solidity/bsh
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bsh.token.js \
-    --method fundBSH --addr $(cat $CONFIG_DIR/token_bsh.proxy.bsc) --amount 1000)
-  echo "$tx" >$CONFIG_DIR/tx/fundBSH.bsc
-}
-
-deposit_token_for_bob() {
-  echo "Funding BOB"
-  cd $CONTRACTS_DIR/solidity/bsh
-  tx=$(truffle exec --network bsc "$SCRIPTS_DIR"/bsh.token.js \
-    --method fundBOB --addr $(get_bob_address) --amount $1)
-  echo "$tx" >$CONFIG_DIR/fundBOB.bsc
-}
-
-token_approveTransfer() {
-  cd $CONTRACTS_DIR/solidity/bsh
-  truffle exec --network bsc "$SCRIPTS_DIR"/bsh.token.js \
-    --method approve --addr $(cat $CONFIG_DIR/token_bsh.proxy.bsc) --amount $1 --from "$(get_bob_address)"
-}
-
-bsc_init_btp_transfer() {
-  ICON_NET=$(cat $CONFIG_DIR/net.btp.icon)
-  ALICE_ADDRESS=$(get_alice_address)
-  BTP_TO="btp://$ICON_NET/$ALICE_ADDRESS"
-  cd $CONTRACTS_DIR/solidity/bsh
-  truffle exec --network bsc "$SCRIPTS_DIR"/bsh.token.js \
-    --method transfer --to $BTP_TO --amount $1 --from "$(get_bob_address)"
-}
-
-calculateTransferFee() {
-  cd $CONTRACTS_DIR/solidity/TokenBSH
-  truffle exec --network bsc "$SCRIPTS_DIR"/bsh.token.js \
-    --method calculateTransferFee --amount $1
-}
-
-get_Bob_Token_Balance() {
-  cd $CONTRACTS_DIR/solidity/bsh
-  BSC_USER=$(get_bob_address)
-  BOB_BALANCE=$(truffle exec --network bsc "$SCRIPTS_DIR"/bsh.token.js \
-    --method getBalance --addr $BSC_USER)
-}
-
-get_Bob_Token_Balance_with_wait() {
-  echo "Checking Bob's Balance after BTP transfer:"
-  get_Bob_Token_Balance
-  BOB_INITIAL_BAL=$BOB_BALANCE
-  COUNTER=30
-  while true; do
-    printf "."
-    if [ $COUNTER -le 0 ]; then
-      printf "\nError: timed out while getting Bob's Balance: Balance unchanged\n"
-      echo $BOB_CURRENT_BAL
-      exit 1
-    fi
-    sleep 3
-    COUNTER=$(expr $COUNTER - 3)
-    get_Bob_Token_Balance
-    BOB_CURRENT_BAL=$BOB_BALANCE
-    if [ "$BOB_CURRENT_BAL" != "$BOB_INITIAL_BAL" ]; then
-      printf "\nBTP Transfer Successfull! \n"
-      break
-    fi
-  done
-  echo "Bob's Balance after BTP transfer: $BOB_CURRENT_BAL ETH"
+  coinId=$(echo "$tx" | grep "coinId:" | sed -e "s/^coinId: //")
+  exists=$(echo $coinId | wc -l)
+  if [ "$exists" != "1" ]; 
+  then
+    echo "Error getting coinID " $1
+    return 1
+  else 
+    echo "$coinId" >$CONFIG_DIR/bsc.addr.coin$1
+  fi 
 }
 
 generate_metadata() {
@@ -218,12 +271,12 @@ generate_metadata() {
     echo "###################  Generating BMC Solidity metadata ###################"
 
     local BMC_ADDRESS=$(jq -r '.networks[] | .address' build/contracts/BMCPeriphery.json)
-    echo btp://$BSC_BMC_NET/"${BMC_ADDRESS}" >$CONFIG_DIR/btp.bsc.btp.address
-    echo "${BMC_ADDRESS}" >$CONFIG_DIR/btp.bsc.bmc.periphery
-    wait_for_file $CONFIG_DIR/btp.bsc.bmc.periphery
+    echo btp://$BSC_BMC_NET/"${BMC_ADDRESS}" >$CONFIG_DIR/bsc.addr.bmcbtp
+    echo "${BMC_ADDRESS}" >$CONFIG_DIR/bsc.addr.bmcperiphery
+    wait_for_file $CONFIG_DIR/bsc.addr.bmcperiphery
 
-    jq -r '.networks[] | .address' build/contracts/BMCManagement.json >$CONFIG_DIR/btp.bsc.bmc.management
-    wait_for_file $CONFIG_DIR/btp.bsc.bmc.management
+    jq -r '.networks[] | .address' build/contracts/BMCManagement.json >$CONFIG_DIR/bsc.addr.bmcmanagement
+    wait_for_file $CONFIG_DIR/bsc.addr.bmcmanagement
     echo "DONE."
     ;;
 
@@ -231,15 +284,10 @@ generate_metadata() {
     BTS)
     echo "################### Generating BTS  Solidity metadata ###################"
 
-    jq -r '.networks[] | .address' build/contracts/BTSCore.json >$CONFIG_DIR/btp.bsc.bts.core
-    wait_for_file $CONFIG_DIR/btp.bsc.bts.core
-    jq -r '.networks[] | .address' build/contracts/BTSPeriphery.json >$CONFIG_DIR/btp.bsc.bts.periphery
-    wait_for_file $CONFIG_DIR/btp.bsc.bts.periphery
-    # jq -r '.networks[] | .address' build/contracts/HRC20.json >$CONFIG_DIR/btp.bsc.tbnb
-    # wait_for_file $CONFIG_DIR/btp.bsc.tbnb
-    # jq -r '.networks[] | .address' build/contracts/ERC20TKN.json >$CONFIG_DIR/btp.bsc.eth
-    # wait_for_file $CONFIG_DIR/btp.bsc.eth
-
+    jq -r '.networks[] | .address' build/contracts/BTSCore.json >$CONFIG_DIR/bsc.addr.btscore
+    wait_for_file $CONFIG_DIR/bsc.addr.btscore
+    jq -r '.networks[] | .address' build/contracts/BTSPeriphery.json >$CONFIG_DIR/bsc.addr.btsperiphery
+    wait_for_file $CONFIG_DIR/bsc.addr.btsperiphery
     echo "DONE."
     ;;
   *)
